+++ changelog.md
@@ -1,3 +1,5 @@
-# ChemLib 1.19.2-2.0.12
+# ChemLib 1.19.2-2.0.13
 
-- Add mob effects to chemicals that can be used later in game to effect the player based on their actions. (Thanks IInventedDirt!)
+Changes:
+- Update Forge, parchment, and bump for parity.
+- Tag all solid chemical block items as storage blocks.